--- src/components/base/Spinner.tsx
@@ -1,7 +1,0 @@
-import { ImSpinner8 } from 'react-icons/im';
-
-const Spinner = () => {
-  return <ImSpinner8 className="spinner" data-testid="spinner-icon" />;
-};
-
-export default Spinner;
