+++ src/components/icons/RemoveButton.tsx
@@ -4,7 +4,7 @@ import ItemButton from './ItemButton';
 const RemoveButton = ({ handleClick }: { handleClick: () => void }) => {
   return (
     <ItemButton
-      icon={<FaTrash className="btn-trash" />}
+      icon={<FaTrash className="btn-trash" data-testid="trash-icon" />}
       handleClick={handleClick}
     />
   );
