+++ src/components/icons/Spinner.test.tsx
@@ -1,6 +1,7 @@
 import { render } from '@testing-library/react';
-import Spinner from './Spinner';
+
 import '@testing-library/jest-dom';
+import Spinner from './Spinner';
 
 describe('Spinner', () => {
   it('renders without crashing', () => {
