--- src/components/todo/RemoveButton.tsx
@@ -1,12 +1,0 @@
-import { FaTrash } from 'react-icons/fa';
-import ItemButton from '../base/ItemButton';
-
-const RemoveButton = ({ handleClick }: { handleClick: () => void }) => {
-  return (
-    <ItemButton handleClick={handleClick}>
-      <FaTrash className="btn-trash" />
-    </ItemButton>
-  );
-};
-
-export default RemoveButton;
