--- src/components/todo/SearchIcon.tsx
@@ -1,11 +1,0 @@
-import { IoSearch } from 'react-icons/io5';
-
-const SearchIcon = () => {
-  return (
-    <div className="input-search" data-testid="test-search-icon">
-      <IoSearch className="btn-search" />
-    </div>
-  );
-};
-
-export default SearchIcon;
