+++ src/components/todo/Todo.css
@@ -59,11 +59,6 @@
   text-overflow: ellipsis;
 }
 
-.btn-plus {
-  color: darkcyan;
-  font-size: 20px;
-}
-
 .btn-search {
   color: #7d7d7d;
   font-size: 20px;
